+++ src/Happstack/Server/SimpleHTTP.hs
@@ -58,8 +58,6 @@ module Happstack.Server.SimpleHTTP
     , module Happstack.Server.Response
     -- ** Request Routing
     , module Happstack.Server.Routing
-    -- ** Proxying
-    , module Happstack.Server.Proxy
     -- ** Looking up values in Query String, Request Body, and Cookies
     , module Happstack.Server.RqData
     -- ** Output Validation
@@ -76,7 +74,6 @@ import Happstack.Server.Monads
 import Happstack.Server.Cookie
 import Happstack.Server.Error
 import Happstack.Server.Types
-import Happstack.Server.Proxy
 import Happstack.Server.Routing
 import Happstack.Server.RqData
 import Happstack.Server.Response
@@ -87,7 +84,7 @@ import Data.Maybe                                (fromMaybe)
 import qualified Data.Version                    as DV
 import Happstack.Server.Internal.Monads          (FilterFun, WebT(..), unFilterFun, runServerPartT, ununWebT)
 import qualified Happstack.Server.Internal.Listen as Listen (listen, listen',listenOn, listenOnIPv4) -- So that we can disambiguate 'Writer.listen'
-import Network                                   (Socket)
+import Network.Socket                            (Socket)
 import qualified Paths_happstack_server          as Cabal
 import System.Console.GetOpt                     ( OptDescr(Option)
                                                  , ArgDescr(ReqArg)