+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+7.6.0
+=====
+
+ - updated to support network 3.* and GHC 8.8.1
+
+ - removed Happstack.Server.Client and Happstack.Server.Proxy under the
+   belief that no one uses them. This makes the upgrade to network 3
+   easier. If you used these, let us know.
+
 7.5.1
 =====